Rename image.translate and move keypoint augmenters to the example
Address naming and layering feedback on the augmentation helpers.

- Rename paz.image.translate_image to paz.image.translate, matching the
  module's own convention (rotate, crop, resize, pad do not repeat "image").
- Move the two-type augmentation recipes (rotate/translate an image together
  with its keypoints) out of paz.keypoints into the keypoints example, where
  they read as user code composing the backend primitives; drop image_center2D.
  The backend keeps only single-type primitives: paz.image.rotate/translate and
  paz.keypoints.rotate_keypoints2D/translate_keypoints.
